requirements.txt
tensorflow-federated==0.19.0
tensorflow-gpu==2.5.0
nest-asyncio
